feat(FR-1229): Migrate BAITable component to backend.ai-ui for File Explorer (#3940)
resolves #3926 (FR-1229)

### Move BAITable component to backend.ai-ui package

This PR moves the BAITable component from the React application to the backend.ai-ui package to make it reusable across the application. The component has been refactored to remove theme-specific code and simplified to focus on core functionality.

Key changes:
- Moved BAITable from `react/src/components/BAITable.tsx` to `packages/backend.ai-ui/src/components/BAITable.tsx`
- Added `react-resizable` as a peer dependency for the resizable table feature
- Added React, React DOM, and Relay as root dependencies
- Updated all imports across the application to use the new component from backend.ai-ui
- Moved the `transformSorterToOrderString` helper function to backend.ai-ui
- Simplified the component by removing the neoStyle prop and ConfigProvider wrapper

**Checklist:**

- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after
