+++ packages/backend.ai-ui/src/components/Table/index.ts
@@ -1,0 +1,2 @@
+export { default as BAITable } from './BAITable';
+export type { BAITableProps } from './BAITable';
