+++ packages/backend.ai-ui/src/components/index.ts
@@ -1,12 +1,11 @@
-// Direct components
 export { default as Flex } from './Flex';
 export type { FlexProps } from './Flex';
 export { default as BAICard } from './BAICard';
 export type { BAICardProps } from './BAICard';
 export { default as BAIResourceWithSteppedProgress } from './BAIResourceWithSteppedProgress';
 export type { BAIResourceWithSteppedProgressProps } from './BAIResourceWithSteppedProgress';
+export * from './Table';
 
-// Subdirectory exports
 export * from './fragments';
 export * from './tests';
 export * from './provider';
