+++ packages/backend.ai-ui/src/locale/vi.json
@@ -2,5 +2,8 @@
   "$schema": "../../i18n.schema.json",
   "comp:BAITestButton": {
     "Test": "kiểm tra"
+  },
+  "comp:PaginationInfoText": {
+    "Total": "{{start}} - {{end}} của {{total}}"
   }
 }
