+++ packages/backend.ai-ui/vite.config.ts
@@ -43,6 +43,11 @@ export default defineConfig(({ mode }) => {
           'react-i18next',
           'relay-runtime',
           'antd',
+          'antd-style',
+          'ahooks',
+          'lodash',
+          'graphql',
+          'classnames',
         ],
       },
       sourcemap: true,
