+++ react/package.json
@@ -55,7 +55,6 @@
     "react-i18next": "^15.4.1",
     "react-markdown": "^10.1.0",
     "react-relay": "^20.1.0",
-    "react-resizable": "^3.0.5",
     "react-router-dom": "^6.30.0",
     "react-scripts": "5.0.1",
     "react-syntax-highlighter": "^15.6.1",
@@ -143,7 +142,6 @@
     "@types/react-copy-to-clipboard": "^5.0.7",
     "@types/react-dom": "^19.0.3",
     "@types/react-relay": "^18.2.1",
-    "@types/react-resizable": "^3.0.8",
     "@types/react-syntax-highlighter": "^15.5.13",
     "@types/react-test-renderer": "^19.0.0",
     "@types/relay-runtime": "^19.0.2",
