+++ react/src/components/CustomizedImageList.tsx
@@ -19,7 +19,6 @@ import {
 } from '../hooks';
 import { useHiddenColumnKeysSetting } from '../hooks/useHiddenColumnKeysSetting';
 import AliasedImageDoubleTags from './AliasedImageDoubleTags';
-import BAITable from './BAITable';
 import { ImageTags } from './ImageTags';
 import TextHighlighter from './TextHighlighter';
 import {
@@ -32,6 +31,7 @@ import { useToggle } from 'ahooks';
 import { App, Button, Input, Popconfirm, theme, Typography } from 'antd';
 import { AnyObject } from 'antd/es/_util/type';
 import { ColumnsType, ColumnType } from 'antd/es/table';
+import { BAITable } from 'backend.ai-ui';
 import _ from 'lodash';
 import React, {
   PropsWithChildren,
@@ -484,8 +484,6 @@ const CustomizedImageList: React.FC<PropsWithChildren> = ({ children }) => {
           />
         </Flex>
         <BAITable
-          neoStyle
-          size="small"
           resizable
           loading={isPendingSearchTransition}
           columns={
