+++ react/src/components/ImageList.tsx
@@ -16,7 +16,6 @@ import {
   useUpdatableState,
 } from '../hooks';
 import { useHiddenColumnKeysSetting } from '../hooks/useHiddenColumnKeysSetting';
-import BAITable from './BAITable';
 import DoubleTag from './DoubleTag';
 import ImageInstallModal from './ImageInstallModal';
 import { ImageTags } from './ImageTags';
@@ -35,6 +34,7 @@ import {
 import { useToggle, useDebounceFn } from 'ahooks';
 import { App, Button, Input, Tag, theme, Tooltip, Typography } from 'antd';
 import { ColumnType } from 'antd/es/table';
+import { BAITable } from 'backend.ai-ui';
 import _ from 'lodash';
 import { Key, useMemo, useState, useTransition } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -545,8 +545,6 @@ const ImageList: React.FC<{ style?: React.CSSProperties }> = ({ style }) => {
           </Flex>
         </Flex>
         <BAITable
-          neoStyle
-          size="small"
           resizable
           rowKey="id"
           scroll={{ x: 'max-content' }}
