+++ react/src/components/SharedFolderPermissionInfoModal.tsx
@@ -8,7 +8,6 @@ import { useCurrentUserInfo } from '../hooks/backendai';
 import { useTanMutation } from '../hooks/reactQueryAlias';
 import { usePainKiller } from '../hooks/usePainKiller';
 import BAIModal, { BAIModalProps } from './BAIModal';
-import BAITable from './BAITable';
 import Flex from './Flex';
 import VFolderPermissionCell from './VFolderPermissionCell';
 import { UserOutlined } from '@ant-design/icons';
@@ -22,7 +21,7 @@ import {
   Typography,
   theme,
 } from 'antd';
-import { BAIUserUnionIcon } from 'backend.ai-ui';
+import { BAITable, BAIUserUnionIcon } from 'backend.ai-ui';
 import { LogOut } from 'lucide-react';
 import { useTranslation } from 'react-i18next';
 import { graphql, useFragment } from 'react-relay';
