+++ react/src/helper/index.tsx
@@ -4,7 +4,6 @@ import { EnvironmentImage } from '../components/ImageList';
 import { useSuspendedBackendaiClient } from '../hooks';
 import { AttachmentsProps } from '@ant-design/x';
 import { fetchEventSource } from '@microsoft/fetch-event-source';
-import { SorterResult } from 'antd/es/table/interface';
 import Big from 'big.js';
 import dayjs from 'dayjs';
 import { Duration } from 'dayjs/plugin/duration';
@@ -408,24 +407,6 @@ export const localeCompare = (a?: string | null, b?: string | null) => {
   if (b === null || b === undefined) return 1;
   return a.localeCompare(b);
 };
-
-export function transformSorterToOrderString<T = any>(
-  sorter: SorterResult<T> | Array<SorterResult<T>>,
-) {
-  if (Array.isArray(sorter)) {
-    return _.chain(sorter)
-      .map((s) =>
-        s.order ? `${s.order === 'descend' ? '-' : ''}${s.field}` : undefined,
-      )
-      .compact()
-      .join(',')
-      .value();
-  } else {
-    return sorter.order
-      ? `${sorter.order === 'descend' ? '-' : ''}${sorter.field}`
-      : undefined;
-  }
-}
 
 export const numberSorterWithInfinityValue = (
   a?: number | null,
