+++ react/src/pages/UserCredentialsPage.tsx
@@ -20,8 +20,8 @@ const useStyles = createStyles(({ css }) => ({
 
 const UserCredentialsPage: React.FC = () => {
   const { t } = useTranslation();
-  const { styles } = useStyles();
   const { token } = theme.useToken();
+  const { styles } = useStyles();
   const [searchParams] = useSearchParams();
   const currentTab = searchParams.get('tab') || 'users';
   const navigate = useNavigate();
