+++ resources/theme.json
@@ -24,6 +24,10 @@
       },
       "Divider": {
         "colorSplit": "#D9D9D9"
+      },
+      "Table": {
+        "headerBg": "#E3E3E3",
+        "headerSplitColor": "rgba(0,0,0,0.25)"
       }
     }
   },
